fix: proper tests for front end
