+++ package.json
@@ -18,7 +18,8 @@
     "eject": "react-native-scripts eject",
     "android": "react-native-scripts android",
     "ios": "react-native-scripts ios",
-    "test": "node node_modules/jest/bin/jest.js --watch",
+    "test": "node node_modules/jest/bin/jest.js",
+    "test2": "mocha tests/utils_tests/*.js",
     "jest": "jest -u"
   },
   "jest": {
